Fix: move file system related functions to the service
Delete:
- readLocalFile.ts
- deleteLocalFile.ts
Update:
- file.service.ts
- file.module.ts
New:
- create File interface
