+++ packages/apiserver/src/file/file.d.ts
@@ -1,0 +1,4 @@
+export interface File {
+  originalname: string;
+  buffer: Buffer;
+}
