+++ packages/helper-functions/src/index.ts
@@ -1,2 +1,1 @@
 export * from './lib/normalizeIdentifier';
-export * from './lib/file-system-services';
