--- packages/helper-functions/src/lib/file-system-services/index.ts
@@ -1,2 +1,0 @@
-export * from './deleteLocalFile';
-export * from './readLocalFile';
